requirements.txt
Flask
cpu-load-generator
psycopg2-binary
python-dotenv